Fix loading of plugins
